Add mocks for a basic course and SF1624
