--- packages/t2l-frontend/src/mocks/handlers.ts
@@ -1,64 +1,0 @@
-import { rest } from "msw";
-import {
-  CanvasGrades,
-  Columns,
-  GradeableStudents,
-  Sections,
-} from "t2l-backend/src/apiHandlers/utils/types";
-import { canvasGrades, ladokGrades } from "./mockData1";
-
-export const handlers = [
-  rest.get(
-    "/transfer-to-ladok/api/courses/mock-0/sections",
-    (req, res, ctx) => {
-      return res(
-        ctx.status(200),
-        ctx.json<Sections>({
-          aktivitetstillfalle: [],
-          kurstillfalle: [],
-        })
-      );
-    }
-  ),
-
-  rest.get(
-    "/transfer-to-ladok/api/courses/mock-1/sections",
-    (req, res, ctx) => {
-      return res(
-        ctx.status(200),
-        ctx.json<Sections>({
-          aktivitetstillfalle: [
-            {
-              id: "0000-000-000",
-              date: "2022-01-01",
-              name: "AKT 01",
-            },
-          ],
-          kurstillfalle: [],
-        })
-      );
-    }
-  ),
-
-  rest.get("/transfer-to-ladok/api/courses/mock-1/columns", (req, res, ctx) =>
-    res(
-      ctx.status(200),
-      ctx.json<Columns>({
-        assignments: [],
-        finalGrades: {
-          hasLetterGrade: true,
-        },
-      })
-    )
-  ),
-
-  rest.get(
-    "/transfer-to-ladok/api/courses/mock-1/ladok-grades",
-    (req, res, ctx) =>
-      res(ctx.status(200), ctx.json<GradeableStudents>(ladokGrades))
-  ),
-
-  rest.get("/transfer-to-ladok/api/courses/mock-1/total", (req, res, ctx) =>
-    res(ctx.status(200), ctx.json<CanvasGrades>(canvasGrades))
-  ),
-];
